move InventoryWindow helpers to inventory, make pub
